requirements.txt
phidata
streamlit
google-generativeai
duckdb
pandas
numpy
agno
openpyxl